backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.5
pydantic-settings==2.7.0

# Environment & Configuration
python-dotenv==1.0.1

# Database
supabase==2.10.0

# NLP & AI
spacy==3.8.3
groq==0.14.0

# Caching & Queue
redis==5.2.1

# Rate Limiting
fastapi-limiter==0.1.6

# Retry Logic
tenacity==9.0.0

# File Upload
python-multipart==0.0.20

# Performance
uvloop==0.21.0
httptools==0.6.4

# Security
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.28.1
